+++ src/app/overall-left/days-left/days-left.component.ts
@@ -1,19 +1,17 @@
 import { Component, OnInit } from '@angular/core';
-import {DaysLeft, DaysLeftService} from "./days-left.service";
+import {TotalLeft, DaysLeftService} from "./days-left.service";
 import {BehaviorSubject} from "rxjs";
 
 @Component({
   selector: 'days-left',
   templateUrl: './days-left.component.html',
-  styleUrls: ['./days-left.component.css']
+  styleUrls: ['../left.css']
 })
 export class DaysLeftComponent implements OnInit {
-  $subject: BehaviorSubject<DaysLeft>;
+  $subject: BehaviorSubject<TotalLeft>;
 
   constructor(private daysLeftService: DaysLeftService) {
-    this.$subject = daysLeftService.$daysLeft;
-
-    this.$subject.subscribe(console.log);
+    this.$subject = daysLeftService.$subject;
   }
 
   ngOnInit(): void {