+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 - basic styles
 - days functionality
+- money functionality
 
 ## [3.2.0] - 2021-12-24
 ### Added